[ui] Move the correlation widgets under fibsem/ui
fibsem/correlation/ui/ imported nine different fibsem.ui modules -- custom_widgets,
icon, tokens, stylesheets, notification_service, utils, qt.threading -- while the
38 non-UI files in fibsem/correlation/ imported none of it. The dependency
already ran one way; only the directory layout disagreed.

Moving it to fibsem/ui/correlation/widgets/ matches the existing
fibsem/ui/fm/widgets/ shape, and leaves fibsem/correlation/ with no Qt import at
all -- it stays the cheap, headless package it already was (44 modules).

The concrete win is the palette guard. test_shared_palette scans anything under
fibsem/ui, plus any file that happens to mention "stylesheets". Six of these
widgets were caught by that second clause; the other eight were not covered at
all, which is where local one-offs like _RMS_NEUTRAL and _UNFITTED_ICON_COLOR
have been accumulating. All fifteen are now scanned by the rule rather than by
coincidence.

Pure move: fifteen files, all recorded as renames so history follows, plus
import-path updates in thirteen callers. No content changed.

The console entry point in pyproject.toml moves with it --
fibsem-correlation-ui now points at fibsem.ui.correlation.widgets.correlation_tab_widget:main
-- and is verified to still resolve.
