+++ fibsem/ui/correlation/widgets/coordinate_list_widget.py
