+++ fibsem/ui/correlation/widgets/image_point_canvas.py
