+++ pyproject.toml
@@ -83,7 +83,7 @@ fibsem-config-ui = "fibsem.ui.widgets.microscope_config_widget:main"
 fibsem-autolamella-ui = "fibsem.applications.autolamella.ui.AutoLamellaMainUI:run_ui"
 fm-ui = "fibsem.ui.FMAcquisitionWidget:main"
 fm-image-viewer-ui = "fibsem.ui.fm.widgets.fm_image_viewer_widget:main"
-fibsem-correlation-ui = "fibsem.correlation.ui.widgets.correlation_tab_widget:main"
+fibsem-correlation-ui = "fibsem.ui.correlation.widgets.correlation_tab_widget:main"
 fibsem-cli = "fibsem.cli:main"
 
 # Recursive by design. A bare `packages = ["fibsem"]` is NOT recursive: it ships
